TractamentDades/requirements.txt
pika
protobuf
google
grpcio
grpcio_tools
